Predict washing machines
